extract todos section
